Revert "hypr: turn display off when locking"
This doesn't work reliably as when suspending as it'd suspend before sleep is
done and then dpms off upon resume. This might have also caused extreme issues
on a desktop machine I've been having. Reverting this makes them go away.

This was a nice-to-have anyways as hypridle would turn the dpms off after a 2
minutes anyways.

This reverts commit 9dd59ef39ee7062733875d77b7c4a6982a09863c.